requirements.txt
requests
beautifulsoup4
rdkit
Pillow
